+++ src/main/java/com/phantomwing/rusticpancakes/block/custom/PancakeBlock.java
@@ -6,8 +6,12 @@ import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.ItemInteractionResult;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.food.FoodProperties;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -26,14 +30,17 @@ import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import net.neoforged.neoforge.common.Tags;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.function.Supplier;
 
 public class PancakeBlock extends Block {
     public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
     public static final Integer MAX_SERVINGS = 6;
     public static final IntegerProperty SERVINGS = IntegerProperty.create("servings", 0, MAX_SERVINGS - 1);
 
-    public final FoodProperties foodProperties;
+    public final Supplier<Item> servingItem;
 
     protected static final VoxelShape PLATE_SHAPE = Block.box(1.0D, 0.0D, 1.0D, 15.0D, 2.0D, 15.0D);
     protected static final VoxelShape[] PANCAKES_SHAPES =  new VoxelShape[]{
@@ -45,15 +52,24 @@ public class PancakeBlock extends Block {
             Shapes.joinUnoptimized(PLATE_SHAPE, Block.box(3.0D, 2.0D, 3.0D, 13.0D, 3.0D, 13.0D), BooleanOp.OR)
     };
 
-    public PancakeBlock(FoodProperties foodProperties, Properties properties) {
+    public PancakeBlock(Supplier<Item> servingItem, Properties properties) {
         super(properties);
 
-        this.foodProperties = foodProperties;
+        this.servingItem = servingItem;
         this.registerDefaultState(this.stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(SERVINGS, 0));
     }
 
     @Override
-    protected @NotNull InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos, Player player, BlockHitResult hitResult) {
+    public @NotNull ItemInteractionResult useItemOn(ItemStack heldStack, @NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand hand, @NotNull BlockHitResult hit) {
+        if (heldStack.is(Tags.Items.TOOLS_SHEAR)) {
+            return takeServing(level, pos, state, player);
+        }
+
+        return ItemInteractionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
+    }
+
+    @Override
+    protected @NotNull InteractionResult useWithoutItem(@NotNull BlockState state, Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull BlockHitResult hitResult) {
         if (level.isClientSide) {
             if (consumeServing(level, pos, state, player).consumesAction()) {
                 return InteractionResult.SUCCESS;
@@ -67,13 +83,32 @@ public class PancakeBlock extends Block {
         return consumeServing(level, pos, state, player);
     }
 
+    protected ItemInteractionResult takeServing(Level level, BlockPos pos, BlockState state, Player player) {
+        // Drop the serving item.
+        Direction direction = player.getDirection().getOpposite();
+        this.spawnItemEntity(level, this.getServingItem(), pos.getX() + 0.5, pos.getY() + 0.3, pos.getZ() + 0.5,
+                direction.getStepX() * 0.15, 0.05, direction.getStepZ() * 0.15);
+
+        // Remove a serving from the block.
+        this.removeServing(level, pos, state);
+
+        // Play a sound, for taking the serving.
+        level.playSound(null, pos, SoundEvents.WOOL_BREAK, SoundSource.PLAYERS, 0.8F, 0.8F);
+
+        return ItemInteractionResult.SUCCESS;
+    }
+
     /**
      * Eats a pancake from the stack, feeding the player.
      */
     protected InteractionResult consumeServing(Level level, BlockPos pos, BlockState state, Player playerIn) {
         if (!playerIn.canEat(false)) {
+            // If the player is full, no interaction is possible.
             return InteractionResult.PASS;
         } else {
+            ItemStack servingStack = this.getServingItem();
+            FoodProperties foodProperties = servingStack.getFoodProperties(playerIn);
+
             // Apply food effect to the player
             if (foodProperties != null) {
                 playerIn.getFoodData().eat(foodProperties);
@@ -84,13 +119,8 @@ public class PancakeBlock extends Block {
                 }
             }
 
-            // Update the block model. If there are no more servings left, destroy the block.
-            int servingsTaken = state.getValue(SERVINGS);
-            if (servingsTaken < MAX_SERVINGS - 1) {
-                level.setBlock(pos, state.setValue(SERVINGS, servingsTaken + 1), MAX_SERVINGS - 1);
-            } else {
-                level.destroyBlock(pos, true);
-            }
+            // Remove a serving from the block.
+            this.removeServing(level, pos, state);
 
             // Play a sound.
             level.playSound(null, pos, SoundEvents.GENERIC_EAT, SoundSource.PLAYERS, 0.8F, 0.8F);
@@ -99,8 +129,28 @@ public class PancakeBlock extends Block {
         }
     }
 
+    /** Update the block model. If there are no more servings left, destroy the block. */
+    private void removeServing(Level level, BlockPos pos, BlockState state) {
+        int servingsTaken = state.getValue(SERVINGS);
+        if (servingsTaken < MAX_SERVINGS - 1) {
+            level.setBlock(pos, state.setValue(SERVINGS, servingsTaken + 1), MAX_SERVINGS - 1);
+        } else {
+            level.destroyBlock(pos, true);
+        }
+    }
+
+    public ItemStack getServingItem() {
+        return new ItemStack(this.servingItem.get());
+    }
+
+    private static void spawnItemEntity(Level level, ItemStack stack, double x, double y, double z, double xMotion, double yMotion, double zMotion) {
+        ItemEntity entity = new ItemEntity(level, x, y, z, stack);
+        entity.setDeltaMovement(xMotion, yMotion, zMotion);
+        level.addFreshEntity(entity);
+    }
+
     @Override
-    public @NotNull VoxelShape getShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
+    public @NotNull VoxelShape getShape(BlockState state, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull CollisionContext context) {
         return PANCAKES_SHAPES[state.getValue(SERVINGS)];
     }
 
@@ -110,12 +160,12 @@ public class PancakeBlock extends Block {
     }
 
     @Override
-    public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, LevelAccessor level, BlockPos currentPos, BlockPos facingPos) {
+    public @NotNull BlockState updateShape(@NotNull BlockState stateIn, @NotNull Direction facing, @NotNull BlockState facingState, @NotNull LevelAccessor level, @NotNull BlockPos currentPos, @NotNull BlockPos facingPos) {
         return facing == Direction.DOWN && !stateIn.canSurvive(level, currentPos) ? Blocks.AIR.defaultBlockState() : super.updateShape(stateIn, facing, facingState, level, currentPos, facingPos);
     }
 
     @Override
-    public boolean canSurvive(BlockState state, LevelReader level, BlockPos pos) {
+    public boolean canSurvive(@NotNull BlockState state, LevelReader level, BlockPos pos) {
         return level.getBlockState(pos.below()).isSolid();
     }
 
@@ -125,17 +175,17 @@ public class PancakeBlock extends Block {
     }
 
     @Override
-    public int getAnalogOutputSignal(BlockState blockState, Level level, BlockPos pos) {
+    public int getAnalogOutputSignal(BlockState blockState, @NotNull Level level, @NotNull BlockPos pos) {
         return blockState.getValue(SERVINGS);
     }
 
     @Override
-    public boolean hasAnalogOutputSignal(BlockState state) {
+    public boolean hasAnalogOutputSignal(@NotNull BlockState state) {
         return true;
     }
 
     @Override
-    public boolean isPathfindable(BlockState state, PathComputationType type) {
+    public boolean isPathfindable(@NotNull BlockState state, @NotNull PathComputationType type) {
         return false;
     }
 }