+++ changelog.md
@@ -1,3 +1,17 @@
+# 1.1.0
+
+### Additions
+- Overhauled the pancake feature:
+    - Added ability to take a single pancake from a pancakes block, with shears
+    - Added new pancake variants: Pancakes (the classic variant), Pumpkin Pancakes
+    - Added Pancake, Honey Pancake, Chocolate Pancake, Vegetable Pancake, Cherry Blossom Pancake, Pumpkin Pancake
+    - Rebalanced pancake recipes and gained effects
+    - Updated all pancake models
+- Added a new cooking ingredient: Syrup! Required for some new items:
+    - Pancakes and Pumpkin Pancakes require Syrup to craft
+- Added tooltips for pancake effects
+- Added Japanese translations (thank you EndilCrafter!)
+
 # 1.0.0
 
 ### Additions